+++ docs/src/theme/Navbar/Layout/index.js
@@ -1,4 +1,4 @@
-import React from "react"
+import React, { useState, useEffect } from "react"
 import clsx from "clsx"
 import { useThemeConfig } from "@docusaurus/theme-common"
 import {
@@ -6,7 +6,10 @@ import {
   useNavbarMobileSidebar,
 } from "@docusaurus/theme-common/internal"
 import NavbarMobileSidebar from "@theme/Navbar/MobileSidebar"
-import styles from "./styles.module.css"
+import styles from "./styles.module.scss"
+import Link from "@docusaurus/Link"
+import NavbarMobileSidebarToggle from "@theme/Navbar/MobileSidebar/Toggle"
+import { useLocation } from "@docusaurus/router"
 
 function NavbarBackdrop(props) {
   return (
@@ -18,31 +21,66 @@ function NavbarBackdrop(props) {
   )
 }
 export default function NavbarLayout({ children }) {
+  const location = useLocation()
+
+  const [scrolledFromTop, setScrolledFromTop] = useState(false)
+
+  useEffect(() => {
+    const onScroll = () => setScrolledFromTop(window.scrollY > 10)
+    window.addEventListener("scroll", onScroll)
+    return () => window.removeEventListener("scroll", onScroll)
+  }, [])
+
   const {
     navbar: { hideOnScroll, style },
   } = useThemeConfig()
   const mobileSidebar = useNavbarMobileSidebar()
   const { navbarRef, isNavbarVisible } = useHideableNavbar(hideOnScroll)
+
   return (
-    <nav
+    <div
       ref={navbarRef}
       className={clsx(
         "navbar",
         "navbar--fixed-top",
-        hideOnScroll && [
-          styles.navbarHideable,
-          !isNavbarVisible && styles.navbarHidden,
-        ],
-        {
-          "navbar--dark": style === "dark",
-          "navbar--primary": style === "primary",
-          "navbar-sidebar--show": mobileSidebar.shown,
-        }
+        { "navbar-sidebar--show": mobileSidebar.shown },
+        styles.header,
+        scrolledFromTop && styles.scrolledFromTop,
+        { [styles.homeNavbar]: location.pathname === "/" }
       )}
     >
-      {children}
+      <div className={styles.left}>
+        <Link to="/">
+          <img
+            alt="Rune Logo"
+            className={styles.logo}
+            src={require("@site/static/img/home/logo.png").default}
+          />
+        </Link>
+      </div>
+      <div className={styles.right}>
+        <Link to="/docs/quick-start" className={styles.menuLink}>
+          Docs
+        </Link>
+        <Link
+          to="/faq"
+          className={clsx(styles.menuLink, styles.hiddenOnMobile)}
+        >
+          FAQ
+        </Link>
+        <a
+          href="https://github.com/rune/rune-games-sdk"
+          target="_blank"
+          rel="noreferrer"
+          className={clsx(styles.menuBtn, styles.hiddenOnMobile)}
+        >
+          <div className={styles.normal}>GitHub</div>
+          <div className={styles.hover}>GitHub</div>
+        </a>
+        {!mobileSidebar.disabled && <NavbarMobileSidebarToggle />}
+      </div>
       <NavbarBackdrop onClick={mobileSidebar.toggle} />
       <NavbarMobileSidebar />
-    </nav>
+    </div>
   )
 }
