--- docs/src/theme/Navbar/Layout/styles.module.css
@@ -1,7 +1,0 @@
-.navbarHideable {
-  transition: transform var(--ifm-transition-fast) ease;
-}
-
-.navbarHidden {
-  transform: translate3d(0, calc(-100% - 2px), 0);
-}
